refactor: 재사용 컴포넌트 DailyWorkerTemplate 분리
